+++ services/workflows-service/scripts/generate-end-user.ts
@@ -1,5 +1,6 @@
 import { faker } from '@faker-js/faker';
 import { Prisma } from '@prisma/client';
+import { StateTag } from '@ballerine/common';
 
 export const assignedTo = faker.image.avatar();
 export const endUserIds = [
@@ -134,6 +135,7 @@ export const generateBusiness = ({
         workflowDefinitionId,
         createdAt: faker.date.recent(2),
         projectId: projectId,
+        tags: [StateTag.MANUAL_REVIEW],
       },
     },
   };
@@ -190,6 +192,7 @@ export const generateEndUser = ({
           createdAt: faker.date.recent(2),
           projectId: projectId,
           parentRuntimeDataId: workflow.parentRuntimeId,
+          tags: [StateTag.MANUAL_REVIEW],
         },
       },
     };