+++ sdks/web-ui-sdk/CHANGELOG.md
@@ -1,5 +1,26 @@
 # web-ui-sdk
 
+## 1.4.6
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.7.5
+
+## 1.4.5
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.7.4
+
+## 1.4.4
+
+### Patch Changes
+
+- Updated dependencies
+  - @ballerine/common@0.7.3
+
 ## 1.4.3
 
 ### Patch Changes