+++ sdks/workflow-node-sdk/CHANGELOG.md
@@ -1,5 +1,23 @@
 # @ballerine/workflow-node-sdk
 
+## 0.5.6
+
+### Patch Changes
+
+- @ballerine/workflow-core@0.5.6
+
+## 0.5.5
+
+### Patch Changes
+
+- @ballerine/workflow-core@0.5.5
+
+## 0.5.4
+
+### Patch Changes
+
+- @ballerine/workflow-core@0.5.4
+
 ## 0.5.3
 
 ### Patch Changes